Add builder.CloseWriter.
